scripts/requirements.txt
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
prophet>=1.1.4
pyyaml>=6.0.1
requests>=2.31.0 